+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.4.15
 
+- Changed Engine to take `w, h` of screen size when initializing.
+- Removed `setup` method to clean that up.
 
 ## 0.4.14